Upgraded to django 1.9, modularized files
