+++ contentcuration/contentcuration/admin.py
@@ -1,6 +1,7 @@
 from django.contrib import admin
 
-from contentcuration.models import Exercise, AssessmentItem
+from contentcuration.models import Exercise, AssessmentItem, License
 
 admin.site.register(Exercise)
 admin.site.register(AssessmentItem)
+admin.site.register(License)
