+++ contentcuration/contentcuration/models.py
@@ -1,24 +1,10 @@
 from django.db import models
 from mptt.models import MPTTModel, TreeForeignKey
 from django.utils.translation import ugettext as _
+from kolibri.content.models import *
 
-
-class Channel(models.Model):
+class Channel(ChannelMetadata):
     """ Permissions come from association with organizations """
-    name = models.CharField(
-        max_length= 100,
-        verbose_name=_("channel name"),
-    )
-    description = models.TextField(
-        max_length = 300,
-        verbose_name=_("channel description"),
-        help_text=_("Description of what a channel contains"),
-    )
-    author = models.CharField(
-        max_length=100,
-        verbose_name=_("channel author"),
-        help_text=_("Channel author can be a person or an organization"),
-    )
     editors = models.ManyToManyField(
         'auth.User',
         verbose_name=_("editors"),
@@ -33,7 +19,6 @@ class Channel(models.Model):
     class Meta:
         verbose_name = _("Channel")
         verbose_name_plural = _("Channels")
-
 
 class TopicTree(models.Model):
     """Base model for all channels"""
@@ -69,7 +54,7 @@ class TopicTree(models.Model):
         verbose_name_plural = _("Topic trees")
 
 
-class Node(MPTTModel):
+class Node(ContentMetadata):
     """
     By default, all nodes have a title and can be used as a topic.
     """
@@ -77,40 +62,14 @@ class Node(MPTTModel):
     created = models.DateTimeField(auto_now_add=True, verbose_name=_("created"))
     modified = models.DateTimeField(auto_now=True, verbose_name=_("modified"))
 
-    parent = TreeForeignKey('self', null=True, blank=True, related_name='children')
-
-    title = models.CharField(
-        max_length=50,
-        verbose_name=_("title"),
-        default=_("Title"),
-        help_text=_("Node title"),
-    )
-
-    description = models.TextField(
-        max_length=200,
-        verbose_name=_("description"),
-        default=_("Description"),
-        help_text=_("Brief description of what is contained in this folder"),
-    )
-
     published = models.BooleanField(
         default=False,
         verbose_name=_("Published"),
         help_text=_("If published, students can access this item"),
     )
 
-    deleted = models.BooleanField(
-        default=False,
-        verbose_name=_("Deleted"),
-        help_text=_(
-            "Indicates that the node has been deleted, and should only "
-            "be retrievable through the admin backend"
-        ),
-    )
-
     sort_order = models.FloatField(
         max_length=50,
-        #unique=True,
         default=0,
         verbose_name=_("sort order"),
         help_text=_("Ascending, lowest number shown first"),
@@ -121,20 +80,6 @@ class Node(MPTTModel):
         blank=True,
         null=True,
         help_text=_("Organization of person who holds the essential rights"),
-    )
-
-    license = models.ForeignKey(
-        'ContentLicense',
-        null=True,
-        verbose_name=_("license"),
-        help_text=_("License under which the work is distributed"),
-    )
-
-    kind = models.CharField(
-        max_length=50,
-        verbose_name=_("kind"),
-        help_text=_("Type of node (topic, video, exercise, etc.)"),
-        default="Topic"
     )
 
     @property
@@ -157,23 +102,7 @@ class Node(MPTTModel):
         verbose_name = _("Topic")
         verbose_name_plural = _("Topics")
         # Do not allow two nodes with the same name on the same level
-        unique_together = ('parent', 'title')
-
-
-
-class ContentLicense(models.Model):
-    name = models.CharField(
-        max_length=255,
-        default=(""),
-        verbose_name=_("name"),
-        help_text=_("Name of license, e.g. 'Creative Commons Share-Alike 2.0'"),
-    )
-    exists = models.BooleanField(
-        default=False,
-        verbose_name=_("license exists"),
-        help_text=_("Tells whether or not a content item is licensed to share"),
-    )
-
+        #unique_together = ('parent', 'title')
 
 # If we decide to subclass Content:
 #
@@ -214,7 +143,18 @@ class ContentLicense(models.Model):
 #     """
 #     Model for Exercise data
 #     """
-
+class ContentLicense(models.Model):
+    name = models.CharField(
+        max_length=255,
+        default=(""),
+        verbose_name=_("name"),
+        help_text=_("Name of license, e.g. 'Creative Commons Share-Alike 2.0'"),
+    )
+    exists = models.BooleanField(
+        default=False,
+        verbose_name=_("license exists"),
+        help_text=_("Tells whether or not a content item is licensed to share"),
+    )
 
 class Exercise(models.Model):
