+++ contentcuration/contentcuration/settings.py
@@ -41,6 +41,7 @@ INSTALLED_APPS = (
     'django.contrib.staticfiles',
     'rest_framework',
     'django_js_reverse',
+    'kolibri.content',
 )
 
 MIDDLEWARE_CLASSES = (
