--- contentcuration/contentcuration/static/js/edit_channel/clipboard/hbtemplates/clipboard_list.handlebars
@@ -1,30 +1,0 @@
-<div id="clipboard" class="pull-right">
-	{{#if collapsed}}
-		<button class="btn btn-default clipboard-toggler clipboard-button" type="button">
-			<span class="glyphicon glyphicon-menu-down" aria-hidden="true"></span>
-			Clipboard
-			<span class="badge"></span>
-		</button>
-	{{else}}
-		<div id="main-clipboard">
-			<div class="clipboard_header">
-				<h4>
-					<span class="clipboard-toggler glyphicon glyphicon-menu-up" aria-hidden="true"></span>
-					Clipboard
-				</h4>
-			</div>
-			<div id="clipboard_content">
-				<div id = "list" class="list_content">
-					<div id="list_background"></div>
-					<ul class="list-unstyled content-list">
-						{{#each content_list}}
-							<li id="clipboard_item_{{id}}" class="clipboard_item"></li>
-						{{else}}
-							<li><em>No items found.</em></li>
-						{{/each}}
-					</ul>
-				</div>
-			</div>
-		</div>
-	{{/if}}
-</div>
