+++ contentcuration/contentcuration/static/js/edit_channel/exercise_creation/hbtemplates/assessment_item_answer_list.handlebars
@@ -1,0 +1,15 @@
+<ul class="list-group">
+<li class="header-item list-group-item">
+	<div class="row">
+		<div class="col-sm-9">
+			Answer Text
+		 </div>
+		 <div class="col-sm-1">
+		 	Correct?
+		 </div>
+		 <div class="col-sm-2">
+		 </div>
+	</div>
+</li>
+</ul>
+<button class="btn btn-success addanswer" title="Add Answer"><span class="glyphicon glyphicon-plus"></span>Add answer</button>
