+++ contentcuration/contentcuration/static/js/edit_channel/exercise_creation/hbtemplates/assessment_item_edit_toolbar_closed.handlebars
@@ -1,0 +1,6 @@
+<button class="btn btn-danger delete" title="Delete">
+	<span class="glyphicon glyphicon-trash"></span>
+</button>
+<button class="btn btn-success" role="button" data-toggle="collapse" data-parent="#accordion" href="#{{ model.id }}" aria-expanded="false" aria-controls="{{ model.id }}" title="Edit">
+	<span class="glyphicon glyphicon-pencil"></span>
+</button>
