+++ contentcuration/contentcuration/static/js/edit_channel/exercise_creation/hbtemplates/assessment_item_free.handlebars
@@ -1,0 +1,48 @@
+{
+    "question": {
+        "content": "{{parse_newlines question}}\n\n[[☃ free-text-input 1]]",
+        "images": {},
+        "widgets": {
+            "free-text-input 1": {
+                "type": "free-text-input",
+                "alignment": "default",
+                "graded": true,
+                "options": {
+                    "answers": [
+                        {
+                            "value": null,
+                            "status": "ungraded",
+                            "message": "",
+                            "answerForms": [],
+                            "strict": false,
+                            "maxError": null
+                        }
+                    ],
+                    "cols": 50,
+                    "rows": 10,
+                    "coefficient": false,
+                    "labelText": ""
+                },
+                "version": {
+                    "major": 0,
+                    "minor": 0
+                }
+            }
+        }
+    },
+    "answerArea": {
+        "type": "multiple",
+        "options": {
+            "content": "",
+            "images": {},
+            "widgets": {}
+        },
+        "calculator": false,
+        "periodicTable": false
+    },
+    "itemDataVersion": {
+        "major": 0,
+        "minor": 1
+    },
+    "hints": []
+}
