+++ contentcuration/contentcuration/static/js/edit_channel/exercise_creation/hbtemplates/editor_view.handlebars
@@ -1,0 +1,1 @@
+{{{ markdown content }}}
