+++ contentcuration/contentcuration/static/js/edit_channel/exercise_creation/hbtemplates/exercise_edit.handlebars
@@ -1,0 +1,44 @@
+<div class="panel panel-primary">
+	<div class="panel-heading">
+		<div class="row">
+			<div class="col-sm-10">
+				<input type="text" class="form-control" id="title" placeholder="Title" value="{{title}}">
+			</div>
+			<div class="col-sm-2">
+				<button class="btn btn-info download"><span class="glyphicon glyphicon-download-alt" title="Download"></span></button>
+				<button class="btn btn-success save"><span class="glyphicon glyphicon-floppy-save" title="Save"></span></button>
+			</div>
+		</div>
+	</div>
+	<div class="panel-body">
+		<textarea class="form-control" id="description" rows="3">{{ description }}</textarea>
+
+		<div class="panel-group" id="accordion" role="tablist" aria-multiselectable="true">
+		</div>
+		<div class="btn-group btn-group-justified" role="group" aria-label="Question Types">
+			<div class="btn-group" role="group">
+				<div class="disabled btn-default">
+					Add New Question:
+				</div>
+			</div>
+			<div class="btn-group" role="group">
+				<button class="btn btn-primary multiplechoice" title="Multiple Choice">
+				<span class="glyphicon glyphicon-plus"></span>
+					Multiple Choice
+				</button>
+			</div>
+			<div class="btn-group" role="group">
+				<button class="btn btn-success truefalse" title="True/False">
+				<span class="glyphicon glyphicon-plus"></span>
+					True/False
+				</button>
+			</div>
+			<div class="btn-group" role="group">
+				<button class="btn btn-info freeresponse" title="Free Response">
+				<span class="glyphicon glyphicon-plus"></span>
+					Free Response
+				</button>
+			</div>
+		</div>
+	</div>
+</div>
