+++ contentcuration/contentcuration/static/js/edit_channel/exercise_creation/hbtemplates/file_upload.handlebars
@@ -1,0 +1,6 @@
+<div id="dropzone">
+</div>
+<span class="btn btn-success fileinput-button dz-clickable">
+        <i class="glyphicon glyphicon-plus"></i>
+        <span>Add file</span>
+</span>
