+++ contentcuration/contentcuration/static/js/edit_channel/uploader/hbtemplates/file_upload.handlebars
@@ -1,0 +1,7 @@
+<div id="dropzone">
+</div>
+<span class="btn btn-default fileinput-button dz-clickable">
+        <i class="glyphicon glyphicon-plus"></i>
+        <span>Add file</span>
+</span>
+<a class="btn btn-success submit_uploaded_files">Submit</a>
