--- contentcuration/contentcuration/static/js/exercise_creation/hbtemplates/assessment_item_answer.handlebars
@@ -1,14 +1,0 @@
-<li class="list-group-item">
-	<div class="row">
-		<div class="col-sm-9">
-		  <div class="answer editable" data-ph="Answer">
-		  </div>
-		 </div>
-		 <div class="col-sm-1">
-		 	<input class="correct" type="checkbox" aria-label="Correct">
-		 </div>
-	 	 <div class="col-sm-2">
-	 	     <div class="answer-toolbar pull-right"></div>
-	 	 </div>
-    </div>
-</li>
