--- contentcuration/contentcuration/static/js/exercise_creation/hbtemplates/assessment_item_answer_toolbar_closed.handlebars
@@ -1,4 +1,0 @@
-<div class="btn-group" role="group" aria-label="Edit and Delete">
-<button class="btn btn-danger delete"><span class="glyphicon glyphicon-trash" title="Delete"></span></button>
-<button class="btn btn-success edit"><span class="glyphicon glyphicon-pencil" title="Edit"></span></button>
-</div>
