--- contentcuration/contentcuration/static/js/exercise_creation/hbtemplates/assessment_item_answer_toolbar_open.handlebars
@@ -1,4 +1,0 @@
-<div class="btn-group" role="group" aria-label="Edit Tools">
-<button class="btn btn-danger delete"><span class="glyphicon glyphicon-trash" title="Delete"></span></button>
-<button class="btn btn-success edit"><span class="glyphicon glyphicon-floppy-save" title="Save"></span></button>
-</div>
