--- contentcuration/contentcuration/static/js/exercise_creation/hbtemplates/assessment_item_edit.handlebars
@@ -1,19 +1,0 @@
-<div class="panel panel-default">
-<span class="badge">{{ number }}</span>
-<div class="panel-heading" role="tab" id="headingOne">
-	<div class="row">
-		<div class="col-sm-9">
-		  <div class="question editable" data-ph="Question">
-		  </div>
-	    </div>
-	    <div class="col-sm-3">
-	    <div class="toolbar pull-right"></div>
-		</div>
-	</div>
-</div>
-<div id="{{ model.id }}" class="panel-collapse collapse" role="tabpanel" aria-labelledby="headingOne">
-  <div class="panel-body">
-  	<div class="answers">
-  	</div>
-  </div>
-</div>
