--- contentcuration/contentcuration/static/js/exercise_creation/hbtemplates/editor.handlebars
@@ -1,12 +1,0 @@
-<div class='editor'></div>
-<div id="toolbar">
-  <span class="ql-format-group">
-	<span class="ql-format-group">
-	<span title="Bold" class="ql-format-button ql-bold"></span>
-	<span class="ql-format-separator"></span>
-	<span title="Italic" class="ql-format-button ql-italic"></span>
-	</span>
-	<span class="ql-format-group">
-	<span title="Image" class="ql-format-button ql-image"></span>
-	</span>
-</div>
