--- contentcuration/contentcuration/static/js/exercise_creation/hbtemplates/editor_view.handlebars
@@ -1,1 +1,0 @@
-{{{ markdown content }}}
