--- contentcuration/contentcuration/static/js/exercise_creation/hbtemplates/file_upload.handlebars
@@ -1,6 +1,0 @@
-<div id="dropzone">
-</div>
-<span class="btn btn-success fileinput-button dz-clickable">
-        <i class="glyphicon glyphicon-plus"></i>
-        <span>Add file</span>
-</span>
