+++ contentcuration/contentcuration/static/less/exercises.less
@@ -2,10 +2,6 @@ nav.exercise_page {
 	text-align: center;
 }
 
-div#dropzone {
-	min-height: 150px;
-}
-
 li.list-group-item.header-item {
 	background-color: #d9edf7;
 }
