+++ contentcuration/contentcuration/static/less/uploader.less
@@ -11,6 +11,18 @@
 	width: @metadata-width;
 }
 
+.modal{
+	z-index: 9999999999999;
+	margin-top:20px;
+
+	.submit_uploaded_files{
+		cursor:pointer;
+	}
+}
+
+div#dropzone {
+	min-height: 150px;
+}
 
 #uploader, #metadata{
 	@uploader-background-color:#D8D8D8;
