+++ contentcuration/contentcuration/urls.py
@@ -19,7 +19,7 @@ from django.contrib import admin
 from django.contrib.auth import views as auth_views
 from rest_framework import routers, viewsets
 from rest_framework.permissions import AllowAny
-from contentcuration.models import *    # TODO: Change this later?
+from contentcuration.models import *
 import serializers
 import views
 
@@ -56,15 +56,14 @@ class AssessmentItemViewSet(BulkModelViewSet):
     queryset = AssessmentItem.objects.all()
     serializer_class = serializers.AssessmentItemSerializer
 
-
-router = routers.DefaultRouter()
+router = routers.DefaultRouter(trailing_slash=False)
 router.register(r'license', LicenseViewSet)
 router.register(r'channel', ChannelViewSet)
 router.register(r'topictree', TopicTreeViewSet)
 router.register(r'node', NodeViewSet)
 router.register(r'exercise', ExerciseViewSet)
 
-bulkrouter = BulkRouter()
+bulkrouter = BulkRouter(trailing_slash=False)
 bulkrouter.register(r'assessmentitem', AssessmentItemViewSet)
 
 urlpatterns = [
@@ -80,7 +79,7 @@ urlpatterns = [
     url(r'^accounts/logout/$', auth_views.logout, {'template_name': 'registration/logout.html'}),
     url(r'^accounts/', include('django.contrib.auth.urls')),
     url(r'^channels/$', views.channel_list, name='channels'),
-    url(r'channels/(?P<channel_id>\w+)', views.channel, name='channel'),
+    url(r'^channels/(?P<channel_id>\w+)', views.channel, name='channel'),
 ]
